Address code quality review on the panel performance PR
Unify the two conflict-detection engines: detect_for_assignment now runs
the same in-memory detection as the full listing with a single synthetic
subject (it is called after the assignment commits, so its own agenda
item is in the context and self-excludes), leaving each conflict shape
with exactly one definition. The per-facilitator overlap query this
orphans is deleted from the repository and its protocol. Foreign-track
attribution collapses to one shared helper for conflicts and preferred-
slot violations, attribution updates use model_copy instead of a field-
by-field rebuild, and participant limits are indexed directly instead of
silently defaulting (the space lookup keeps its guard: that invariant is
enforced at assign time, not by the database, and one stale row should
skip a capacity warning, not 500 the grid).

The nav-progress fade timers become closures over the detached bar — the
shared hideTimer let a quick follow-up request cancel a finished bar's
cleanup and leave it stuck at full width until the next full page load.

ZEAL_RAISE is now an explicit, schema-declared env flag set by .env.test
and the test:postgres task instead of sniffing pytest's PYTEST_VERSION.
revertible_log_pks requires the logs it dedupes against, dropping the
lazily-refetching optional-parameter path nothing used. The timetable
service tests move to tests/unit/test_timetable_mills.py to mirror the
module move, with new coverage for the unified assignment path.
