+++ .env.schema
@@ -97,6 +97,9 @@ DEBUG_TOOLBAR=false
 IN_TESTS=false
 # @type=boolean @optional
 IS_STAGING=false
+# django-zeal N+1 findings raise (pytest) instead of logging (e2e server).
+# @type=boolean @optional
+ZEAL_RAISE=false
 
 DJANGO_SETTINGS_MODULE=ludamus.edges.settings
 PYTHONPATH=src
