+++ .env.test
@@ -6,6 +6,7 @@ CREDENTIALS_ENCRYPTION_KEY="gWc408f8j3ZQUm5Ly4EXQC2EwFRf50AhePkmZKL5N9E="
 DB_NAME=":memory:"
 ENV="test"
 IN_TESTS='True'
+ZEAL_RAISE='True'
 MEMBERSHIP_API_BASE_URL="https://membership-test.example.com/api/v1/endpoint"
 MEMBERSHIP_API_TOKEN="membership-test-token"
 # Tests drive DBOS explicitly (test_dbos_scheduler.py); never launch implicitly.
