+++ mise.toml
@@ -288,6 +288,7 @@ env.SECRET_KEY = "dev-secret-key"
 env.SESSION_COOKIE_DOMAIN = ".testserver"
 env.CREDENTIALS_ENCRYPTION_KEY = "gWc408f8j3ZQUm5Ly4EXQC2EwFRf50AhePkmZKL5N9E="
 env.IN_TESTS = 'True'
+env.ZEAL_RAISE = 'True'
 # Tests drive DBOS explicitly (test_dbos_scheduler.py); never launch implicitly.
 env.SCHEDULER_MODE = "cron"
