+++ src/ludamus/gates/web/django/chronology/panel/views/timetable.py
@@ -350,7 +350,7 @@ class TimetableAssignView(PanelAccessMixin, EventContextMixin, View):
         self.request.services.waitlist_promotion.fill_freed_seats(session_id=session_pk)
 
         conflicts = ConflictDetectionService(uow).detect_for_assignment(
-            session_pk=session_pk, placement=placement
+            event_pk=current_event.pk, session_pk=session_pk, placement=placement
         )
 
         trigger_data: dict[str, object] = {"timetableChanged": {}}
