+++ src/ludamus/pacts/chronology.py
@@ -220,7 +220,7 @@ class SessionContentEditServiceProtocol(Protocol):
     def list_log(self, event_id: int) -> list[ContentChangeLogDTO]: ...
     def list_field_names(self, event_id: int) -> dict[int, str]: ...
     def revertible_log_pks(
-        self, event_id: int, logs: list[ContentChangeLogDTO] | None = None
+        self, event_id: int, logs: list[ContentChangeLogDTO]
     ) -> set[int]: ...
